# BUG: Fix simd loadable stride logic 