--- a/numpy/_core/src/common/simd/simd.h
+++ b/numpy/_core/src/common/simd/simd.h
@@ -123,10 +123,11 @@ typedef double     npyv_lanetype_f64;
  * acceptable limit of strides before using any of non-contiguous load/store intrinsics.
  *
  * For instance:
- *  npy_intp ld_stride = step[0] / sizeof(float);
- *  npy_intp st_stride = step[1] / sizeof(float);
  *
- *  if (npyv_loadable_stride_f32(ld_stride) && npyv_storable_stride_f32(st_stride)) {
+ *  if (npyv_loadable_stride_f32(steps[0]) && npyv_storable_stride_f32(steps[1])) {
+ *      // Strides are now guaranteed to be a multiple and compatible
+ *      npy_intp ld_stride = steps[0] / sizeof(float);
+ *      npy_intp st_stride = steps[1] / sizeof(float);
  *      for (;;)
  *          npyv_f32 a = npyv_loadn_f32(ld_pointer, ld_stride);
  *          // ...
@@ -134,7 +135,7 @@ typedef double     npyv_lanetype_f64;
  *  }
  *  else {
  *      for (;;)
- *          // C scalars
+ *          // C scalars, use byte steps/strides.
  *  }
  */
 #ifndef NPY_SIMD_MAXLOAD_STRIDE32
@@ -149,11 +150,29 @@ typedef double     npyv_lanetype_f64;
 #ifndef NPY_SIMD_MAXSTORE_STRIDE64
     #define NPY_SIMD_MAXSTORE_STRIDE64 0
 #endif
-#define NPYV_IMPL_MAXSTRIDE(SFX, MAXLOAD, MAXSTORE) \
-    NPY_FINLINE int npyv_loadable_stride_##SFX(npy_intp stride) \
-    { return MAXLOAD > 0 ? llabs(stride) <= MAXLOAD : 1; } \
-    NPY_FINLINE int npyv_storable_stride_##SFX(npy_intp stride) \
-    { return MAXSTORE > 0 ? llabs(stride) <= MAXSTORE : 1; }
+#define NPYV_IMPL_MAXSTRIDE(SFX, MAXLOAD, MAXSTORE)                          \
+    NPY_FINLINE int                                                          \
+    npyv_loadable_stride_##SFX(npy_intp stride)                              \
+    {                                                                        \
+        if (_Alignof(npyv_lanetype_##SFX) != sizeof(npyv_lanetype_##SFX) &&  \
+                stride % sizeof(npyv_lanetype_##SFX) != 0) {                 \
+            /* stride not a multiple of itemsize, cannot handle. */          \
+            return 0;                                                        \
+        }                                                                    \
+        stride = stride / sizeof(npyv_lanetype_##SFX);                       \
+        return MAXLOAD > 0 ? llabs(stride) <= MAXLOAD : 1;                   \
+    }                                                                        \
+    NPY_FINLINE int                                                          \
+    npyv_storable_stride_##SFX(npy_intp stride)                              \
+    {                                                                        \
+        if (_Alignof(npyv_lanetype_##SFX) != sizeof(npyv_lanetype_##SFX) &&  \
+                stride % sizeof(npyv_lanetype_##SFX) != 0) {                 \
+            /* stride not a multiple of itemsize, cannot handle. */          \
+            return 0;                                                        \
+        }                                                                    \
+        stride = stride / sizeof(npyv_lanetype_##SFX);                       \
+        return MAXSTORE > 0 ? llabs(stride) <= MAXSTORE : 1;                 \
+    }
 #if NPY_SIMD
     NPYV_IMPL_MAXSTRIDE(u32, NPY_SIMD_MAXLOAD_STRIDE32, NPY_SIMD_MAXSTORE_STRIDE32)
     NPYV_IMPL_MAXSTRIDE(s32, NPY_SIMD_MAXLOAD_STRIDE32, NPY_SIMD_MAXSTORE_STRIDE32)