--- a/numpy/_core/src/umath/loops_exponent_log.dispatch.c.src
+++ b/numpy/_core/src/umath/loops_exponent_log.dispatch.c.src
@@ -1315,16 +1315,16 @@ NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(DOUBLE_@func@)
 (char **args, npy_intp const *dimensions, npy_intp const *steps, void *NPY_UNUSED(data))
 {
 #if NPY_SIMD && defined(NPY_HAVE_AVX512_SKX) && defined(NPY_CAN_LINK_SVML)
-    const npy_double *src = (npy_double*)args[0];
-    npy_double *dst = (npy_double*)args[1];
-    const int lsize = sizeof(src[0]);
-    const npy_intp ssrc = steps[0] / lsize;
-    const npy_intp sdst = steps[1] / lsize;
     const npy_intp len = dimensions[0];
-    assert(steps[0] % lsize == 0 && steps[1] % lsize == 0);
-    if (!is_mem_overlap(src, steps[0], dst, steps[1], len) &&
-            npyv_loadable_stride_f64(ssrc) &&
-            npyv_storable_stride_f64(sdst)) {
+
+    if (!is_mem_overlap(args[0], steps[0], args[1], steps[1], len) &&
+            npyv_loadable_stride_f64(steps[0]) &&
+            npyv_storable_stride_f64(steps[1])) {
+        const npy_double *src = (npy_double*)args[0];
+        npy_double *dst = (npy_double*)args[1];
+        const npy_intp ssrc = steps[0] / sizeof(src[0]);
+        const npy_intp sdst = steps[1] / sizeof(src[0]);
+
         simd_@func@_f64(src, ssrc, dst, sdst, len);
         return;
     }