--- a/numpy/_core/src/umath/loops_unary.dispatch.c.src
+++ b/numpy/_core/src/umath/loops_unary.dispatch.c.src
@@ -298,12 +298,12 @@ NPY_NO_EXPORT void NPY_CPU_DISPATCH_CURFX(@TYPE@_@kind@)
             goto clear;
         }
     #if @supports_ncontig@
-        const npy_intp istride = istep / sizeof(STYPE);
-        const npy_intp ostride = ostep / sizeof(STYPE);
-        if (TO_SIMD_SFX(npyv_loadable_stride)(istride) &&
-            TO_SIMD_SFX(npyv_storable_stride)(ostride))
+        if (TO_SIMD_SFX(npyv_loadable_stride)(istep) &&
+            TO_SIMD_SFX(npyv_storable_stride)(ostep))
         {
-            if (istride == 1 && ostride != 1) {
+            const npy_intp istride = istep / sizeof(STYPE);
+            const npy_intp ostride = ostep / sizeof(STYPE);
+            if (istride == sizeof(STYPE) && ostride != 1) {
                 // contiguous input, non-contiguous output
                 TO_SIMD_SFX(simd_unary_cn_@intrin@)(
                     (STYPE*)ip, (STYPE*)op, ostride, len