@seberg 2024-07-24
I have confirmed that this fixes the issue in the i386 docker env.  It should be carefully reviewed anyway, since I am not sure how good the tests really are.  The tests at gh-27028 crashed (presumably hitting one of the deleted asserts).