@charris 2024-07-26
Does this need a backport? There are a couple of other SIMD fixes...