@mattip 2024-07-30
> After removing the asserts, they failed beautifully.

nice