@mattip 2024-07-30
> I assume that compilers will optimize the duplicate division away,

The code looks good to me. Could you benchmark a relevant ufunc to see if there is any impact?