@r-devulap 2024-07-30
> the exp implementation has it's own custom logic, that is probably OK.

You mean here? https://github.com/numpy/numpy/blob/893215af7f97105fa2dccb5bb560bc322dda6c77/numpy/_core/src/umath/loops_exponent_log.dispatch.c.src#L1287 Perhaps, we should use the same logic everywhere for consistency?