@seberg 2024-07-31
Not sure there is a test that really might notice it (unless we disable simd accidentally), but I see these timings.  That is huge fluctuations, but I can't say any of them look systematic (locally on mac).

The only one that seems a bit plausible might be the 3rd one:
```
bench_ufunc_strides.UnaryIntContig.time_unary(<ufunc 'absolute'>, 1, 1, 'Q')
```

<details>

```
| Change   | Before [50bd355e] <main>   | After [e27f4190] <sane-simd-steps>   |   Ratio | Benchmark (Parameter)                                                                           |
|----------|----------------------------|--------------------------------------|---------|-------------------------------------------------------------------------------------------------|
| +        | 2.09±0.01μs                | 3.55±0.06μs                          |    1.7  | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'negative'>, 1, 2, 'f')                    |
| +        | 2.11±0.03μs                | 3.52±0.01μs                          |    1.67 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'negative'>, 1, 2, 'f')                           |
| +        | 12.1±1μs                   | 17.0±0.02μs                          |    1.4  | bench_ufunc_strides.UnaryIntContig.time_unary(<ufunc 'absolute'>, 1, 1, 'Q')                    |
| +        | 9.59±0.01μs                | 12.7±0.1μs                           |    1.33 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 1, 1, 'd')                            |
| +        | 9.81±0.09μs                | 13.0±0.2μs                           |    1.33 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'deg2rad'>, 1, 2, 'd')                     |
| +        | 9.59±0.01μs                | 12.7±0.02μs                          |    1.32 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 1, 1, 'd')                            |
| +        | 9.70±0.01μs                | 12.8±0.04μs                          |    1.32 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 1, 2, 'd')                            |
| +        | 9.70±0μs                   | 12.8±0μs                             |    1.32 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 4, 2, 'd')                            |
| +        | 9.70±0.01μs                | 12.8±0.01μs                          |    1.32 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 1, 2, 'd')                            |
| +        | 9.71±0.03μs                | 12.8±0.02μs                          |    1.32 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 4, 1, 'd')                            |
| +        | 9.71±0.01μs                | 12.8±0.01μs                          |    1.32 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 4, 2, 'd')                            |
| +        | 9.69±0.06μs                | 12.8±0.1μs                           |    1.32 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'deg2rad'>, 1, 1, 'd')                     |
| +        | 9.73±0.01μs                | 12.8±0.01μs                          |    1.32 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'deg2rad'>, 4, 1, 'd')                     |
| +        | 9.62±0.01μs                | 12.7±0μs                             |    1.32 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'radians'>, 1, 1, 'd')                     |
| +        | 9.80±0.1μs                 | 12.8±0.01μs                          |    1.31 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'radians'>, 1, 2, 'd')                     |
| +        | 9.78±0.03μs                | 12.8±0.01μs                          |    1.31 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'radians'>, 4, 2, 'd')                     |
| +        | 9.88±0.2μs                 | 12.9±0.1μs                           |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 4, 1, 'd')                            |
| +        | 9.84±0.09μs                | 12.8±0.1μs                           |    1.3  | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'deg2rad'>, 4, 2, 'd')                     |
| +        | 9.88±0.09μs                | 12.8±0.02μs                          |    1.3  | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'radians'>, 4, 1, 'd')                     |
| +        | 2.15±0.02μs                | 2.72±0.03μs                          |    1.27 | bench_ufunc.CustomComparison.time_less_than_scalar2(<class 'numpy.int16'>)                      |
| +        | 3.71±0.03μs                | 4.41±0μs                             |    1.19 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'reciprocal'>, 4, 1, 'd')                  |
| +        | 2.95±0.1μs                 | 3.51±0μs                             |    1.19 | bench_ufunc_strides.UnaryIntContig.time_unary(<ufunc 'logical_not'>, 1, 1, 'h')                 |
| +        | 3.79±0.3μs                 | 4.47±0.09μs                          |    1.18 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'rint'>, 4, 1, 'd')                        |
| +        | 4.90±0.7μs                 | 5.68±0.01μs                          |    1.16 | bench_ufunc_strides.BinaryComplex.time_binary_scalar_in1(<ufunc 'add'>, 2, 4, 1, 'D')           |
| +        | 3.51±0.4μs                 | 4.04±0.02μs                          |    1.15 | bench_ufunc_strides.BinaryComplex.time_binary_scalar_in1(<ufunc 'add'>, 1, 1, 1, 'D')           |
| +        | 3.55±0.5μs                 | 4.09±0.01μs                          |    1.15 | bench_ufunc_strides.BinaryComplex.time_binary_scalar_in1(<ufunc 'subtract'>, 1, 1, 1, 'D')      |
| +        | 66.2±0.7μs                 | 76.0±4μs                             |    1.15 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'arccos'>, 1, 1, 'd')                             |
| +        | 1.97±0.02μs                | 2.27±0.3μs                           |    1.15 | bench_ufunc_strides.UnaryIntContig.time_unary(<ufunc 'isfinite'>, 1, 1, 'i')                    |
| +        | 3.69±0.03μs                | 4.21±0.04μs                          |    1.14 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 1, 2, 'd')                      |
| +        | 3.59±0.4μs                 | 4.04±0.02μs                          |    1.13 | bench_ufunc_strides.BinaryComplex.time_binary_scalar_in1(<ufunc 'subtract'>, 1, 2, 1, 'D')      |
| +        | 3.69±0.01μs                | 4.18±0.01μs                          |    1.13 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 1, 2, 'd')                           |
| +        | 3.55±0μs                   | 3.98±0μs                             |    1.12 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 4, 2, 'f')                      |
| +        | 3.71±0.01μs                | 4.16±0μs                             |    1.12 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (1), 1, 2, 'd')                      |
| +        | 3.55±0μs                   | 3.98±0.2μs                           |    1.12 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'conjugate'> (1), 4, 2, 'f')               |
| +        | 15.3±1μs                   | 17.1±0.1μs                           |    1.12 | bench_ufunc_strides.UnaryIntContig.time_unary(<ufunc 'sign'>, 1, 1, 'l')                        |
| +        | 3.79±0.03μs                | 4.17±0.04μs                          |    1.1  | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'conjugate'> (1), 1, 2, 'd')               |
| +        | 110±0.01μs                 | 120±0.1μs                            |    1.09 | bench_ufunc.UFunc.time_ufunc_types('radians')                                                   |
| +        | 4.65±0.4μs                 | 5.07±0.02μs                          |    1.09 | bench_ufunc_strides.BinaryComplex.time_binary_scalar_in1(<ufunc 'multiply'>, 4, 4, 1, 'F')      |
| +        | 1.03±0.01μs                | 1.12±0.01μs                          |    1.09 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in1(<ufunc 'multiply'>, 1, 1, 1, 'I')    |
| +        | 67.9±0.8μs                 | 73.9±3μs                             |    1.09 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'arccos'>, 4, 2, 'd')                             |
| +        | 1.33±0.02μs                | 1.46±0.03μs                          |    1.09 | bench_ufunc_strides.UnaryIntContig.time_unary(<ufunc 'isnan'>, 1, 1, 'L')                       |
| +        | 111±1μs                    | 120±0.2μs                            |    1.08 | bench_ufunc.UFunc.time_ufunc_types('deg2rad')                                                   |
| +        | 708±7ns                    | 763±30ns                             |    1.08 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'bitwise_xor'>, 1, 1, 1, 'B') |
| +        | 1.51±0.02μs                | 1.61±0.02μs                          |    1.07 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'add'>, 1, 1, 1, 'Q')         |
| +        | 1.50±0.01μs                | 1.60±0.05μs                          |    1.07 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'bitwise_or'>, 1, 1, 1, 'Q')  |
| +        | 1.50±0.01μs                | 1.60±0.02μs                          |    1.07 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'subtract'>, 1, 1, 1, 'Q')    |
| +        | 851±3ns                    | 910±20ns                             |    1.07 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in1(<ufunc 'add'>, 1, 1, 1, 'h')         |
| +        | 1.50±0.02μs                | 1.60±0.02μs                          |    1.07 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in1(<ufunc 'add'>, 1, 1, 1, 'q')         |
| +        | 3.50±0.04μs                | 3.73±0.04μs                          |    1.07 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'negative'>, 1, 2, 'd')                           |
| +        | 3.50±0.03μs                | 3.76±0.05μs                          |    1.07 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'conjugate'> (0), 4, 1, 'f')               |
| +        | 9.03±0.03μs                | 9.61±0.1μs                           |    1.06 | bench_ufunc.UFunc.time_ufunc_types('bitwise_not')                                               |
| +        | 1.02±0μs                   | 1.08±0.01μs                          |    1.06 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'bitwise_xor'>, 1, 1, 1, 'I') |
| +        | 709±2ns                    | 749±10ns                             |    1.06 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'bitwise_xor'>, 1, 1, 1, 'b') |
| +        | 738±10ns                   | 779±20ns                             |    1.06 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in1(<ufunc 'add'>, 1, 1, 1, 'B')         |
| +        | 1.04±0.01μs                | 1.10±0.02μs                          |    1.06 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in1(<ufunc 'bitwise_and'>, 1, 1, 1, 'i') |
| +        | 794±1ns                    | 838±10ns                             |    1.06 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in1(<ufunc 'logical_and'>, 1, 1, 1, 'H') |
| +        | 739±7ns                    | 784±30ns                             |    1.06 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'floor'>, 1, 1, 'f')                              |
| +        | 3.52±0.01μs                | 3.74±0μs                             |    1.06 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'conjugate'> (1), 4, 1, 'f')               |
| +        | 16.0±0.9μs                 | 17.0±0.01μs                          |    1.06 | bench_ufunc_strides.UnaryIntContig.time_unary(<ufunc 'positive'>, 1, 1, 'q')                    |
| +        | 58.5±0.4μs                 | 61.7±0.2μs                           |    1.05 | bench_ufunc.UFunc.time_ufunc_types('isfinite')                                                  |
| +        | 1.55±0.01μs                | 1.63±0.05μs                          |    1.05 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'bitwise_and'>, 1, 1, 1, 'Q') |
| +        | 701±4ns                    | 739±6ns                              |    1.05 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'bitwise_and'>, 1, 1, 1, 'b') |
| +        | 1.06±0.01μs                | 1.12±0.01μs                          |    1.05 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in0(<ufunc 'bitwise_and'>, 1, 1, 1, 'i') |
| +        | 713±4ns                    | 749±10ns                             |    1.05 | bench_ufunc_strides.BinaryIntContig.time_binary_scalar_in1(<ufunc 'subtract'>, 1, 1, 1, 'b')    |
| +        | 3.55±0.04μs                | 3.74±0.01μs                          |    1.05 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 4, 1, 'f')                      |
| +        | 38.8±0.1μs                 | 40.9±0.6μs                           |    1.05 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'log'>, 4, 2, 'd')                                |
| +        | 69.2±0.2μs                 | 72.9±1μs                             |    1.05 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'log'>, 1, 1, 'e')                         |
| -        | 390±6ns                    | 371±5ns                              |    0.95 | bench_ufunc.ArgParsing.time_add_arg_parsing((array(1.), array(2.)))                             |
| -        | 412±20ns                   | 393±1ns                              |    0.95 | bench_ufunc.CustomArrayFloorDivideInt.time_floor_divide_int(<class 'numpy.uint32'>, 100)        |
| -        | 325±6ns                    | 308±0.8ns                            |    0.95 | bench_ufunc.Methods0DInvert.time_ndarray__0d__('int16')                                         |
| -        | 458±7ns                    | 433±2ns                              |    0.95 | bench_ufunc.MethodsV1.time_ndarray_meth('__le__', 'complex64')                                  |
| -        | 387±7ns                    | 368±0.8ns                            |    0.95 | bench_ufunc.MethodsV1.time_ndarray_meth('__lt__', 'int64')                                      |
| -        | 687±5ns                    | 653±10ns                             |    0.95 | bench_ufunc.MethodsV1.time_ndarray_meth('__matmul__', 'complex64')                              |
| -        | 70.1±2μs                   | 66.7±0.2μs                           |    0.95 | bench_ufunc.UFunc.time_ufunc_types('copysign')                                                  |
| -        | 139±2μs                    | 132±0.2μs                            |    0.95 | bench_ufunc.UFunc.time_ufunc_types('spacing')                                                   |
| -        | 481±10ns                   | 456±5ns                              |    0.95 | bench_ufunc.UFuncSmall.time_ufunc_small_int_array('sqrt')                                       |
| -        | 6.67±0.1μs                 | 6.32±0.02μs                          |    0.95 | bench_ufunc_strides.BinaryComplex.time_binary(<ufunc 'add'>, 4, 1, 2, 'F')                      |
| -        | 7.66±0.2μs                 | 7.28±0.01μs                          |    0.95 | bench_ufunc_strides.BinaryComplex.time_binary_scalar_in0(<ufunc 'multiply'>, 1, 2, 2, 'D')      |
| -        | 3.36±0.01μs                | 3.19±0.2μs                           |    0.95 | bench_ufunc_strides.BinaryComplex.time_binary_scalar_in1(<ufunc 'subtract'>, 2, 2, 1, 'F')      |
| -        | 1.17±0.03μs                | 1.12±0.01μs                          |    0.95 | bench_ufunc_strides.BinaryFPSpecial.time_binary(<ufunc 'maximum'>, 1, 1, 1, 'f')                |
| -        | 479±20ns                   | 453±0.5ns                            |    0.95 | bench_ufunc_strides.BinaryIntContig.time_binary(<ufunc 'bitwise_xor'>, 1, 1, 1, 'b')            |
| -        | 3.73±0.01μs                | 3.54±0.02μs                          |    0.95 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'logical_not'>, 4, 1, 'f')                 |
| -        | 1.61±0.04ms                | 1.51±0.01ms                          |    0.94 | bench_ufunc.CustomArrayFloorDivideInt.time_floor_divide_int(<class 'numpy.int64'>, 1000000)     |
| -        | 374±4ns                    | 353±3ns                              |    0.94 | bench_ufunc.MethodsV1.time_ndarray_meth('__le__', 'int16')                                      |
| -        | 671±20ns                   | 630±4ns                              |    0.94 | bench_ufunc.MethodsV1.time_ndarray_meth('__matmul__', 'float32')                                |
| -        | 438±10ns                   | 411±0.8ns                            |    0.94 | bench_ufunc.MethodsV1.time_ndarray_meth('__ne__', 'complex64')                                  |
| -        | 65.7±4μs                   | 61.5±0.2μs                           |    0.94 | bench_ufunc.UFunc.time_ufunc_types('trunc')                                                     |
| -        | 452±9ns                    | 424±1ns                              |    0.94 | bench_ufunc.UFuncSmall.time_ufunc_numpy_scalar('cos')                                           |
| -        | 321±5ns                    | 303±1ns                              |    0.94 | bench_ufunc.UFuncSmall.time_ufunc_small_array('abs')                                            |
| -        | 327±9ns                    | 308±0.5ns                            |    0.94 | bench_ufunc.UFuncSmall.time_ufunc_small_array('cos')                                            |
| -        | 484±3ns                    | 454±3ns                              |    0.94 | bench_ufunc.UFuncSmall.time_ufunc_small_int_array('cos')                                        |
| -        | 8.89±0.05μs                | 8.38±0μs                             |    0.94 | bench_ufunc_strides.BinaryComplex.time_binary(<ufunc 'add'>, 4, 2, 4, 'F')                      |
| -        | 20.7±0.7μs                 | 19.5±0.04μs                          |    0.94 | bench_ufunc_strides.BinaryComplex.time_binary(<ufunc 'add'>, 4, 4, 4, 'D')                      |
| -        | 14.3±0.05μs                | 13.5±0.03μs                          |    0.94 | bench_ufunc_strides.BinaryComplex.time_binary_scalar_in1(<ufunc 'add'>, 4, 4, 4, 'D')           |
| -        | 7.90±0.06μs                | 7.46±0.01μs                          |    0.94 | bench_ufunc_strides.BinaryFP.time_binary_scalar_in0(<ufunc 'fmin'>, 2, 4, 4, 'd')               |
| -        | 3.74±0.02μs                | 3.53±0.01μs                          |    0.94 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'logical_not'>, 4, 1, 'f')                        |
| -        | 1.61±0.1μs                 | 1.51±0.02μs                          |    0.94 | bench_ufunc_strides.UnaryIntContig.time_unary(<ufunc 'conjugate'>, 1, 1, 'b')                   |
| -        | 1.41±0.04μs                | 1.31±0.02μs                          |    0.93 | bench_ufunc.CustomScalarFloorDivideUInt.time_floor_divide_uint(<class 'numpy.uint16'>, 43)      |
| -        | 320±5ns                    | 298±0.9ns                            |    0.93 | bench_ufunc.Methods0DBoolComplex.time_ndarray__0d__('__complex__', 'float32')                   |
| -        | 211±4μs                    | 197±1μs                              |    0.93 | bench_ufunc.UFunc.time_ufunc_types('divide')                                                    |
| -        | 15.5±0.07μs                | 14.4±0.8μs                           |    0.93 | bench_ufunc_strides.BinaryComplex.time_binary(<ufunc 'divide'>, 4, 4, 1, 'D')                   |
…